Moved db stuff to pkg
