+++ pkg/database/database_errors.go
