+++ pkg/database/mysql/anonymiser_test.go
@@ -4,8 +4,8 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/hellofresh/klepto/database"
-	"github.com/hellofresh/klepto/utils"
+	"github.com/hellofresh/klepto/pkg/database"
+	"github.com/hellofresh/klepto/pkg/utils"
 )
 
 var v interface{}
